projector/requirements.txt
# Core dependencies
flask==2.0.1
python-dotenv==0.19.0
requests==2.26.0
slack-sdk==3.11.2
PyGithub==1.55
streamlit==1.3.0
pydantic==1.9.0
langchain==0.0.267
langchain-core==0.1.0
unidiff==0.7.4

# Agentgen and Codegen dependencies
agentgen>=0.1.0
git+https://github.com/Zeeeepa/codegen.git
codeowners==0.6.0
gitpython==3.1.30

# Backend dependencies
slack_sdk>=3.19.0
PyGithub>=1.58.0
markdown>=3.4.0
beautifulsoup4>=4.11.0
python-dotenv>=0.21.0

# Frontend dependencies
streamlit>=1.22.0
streamlit-ace>=0.1.1
streamlit-chat>=0.0.2.2
streamlit-toggle-switch>=1.0.2
plotly>=5.10.0
pandas>=1.5.3
matplotlib>=3.7.1
plotly>=5.14.1
pydantic>=1.10.7
requests>=2.28.2
aiohttp>=3.8.4
asyncio>=3.4.3
typing-extensions>=4.5.0
tqdm>=4.65.0
colorama>=0.4.6
markdown>=3.4.3
beautifulsoup4>=4.12.2
lxml>=4.9.2
agentgen>=0.1.0
git+https://github.com/Zeeeepa/codegen.git